Migrate Mutation.deleteProject
